+++ src/components/molecules/ListItemDropDown/type.ts
@@ -1,0 +1,2 @@
+export type ListItemDropDownProps = {
+}
